Refactor header and participant preview banner
Both are now defined in a sticky <page-header> element.
