+++ lit/src/components/sidenav/sidenav.scss
@@ -39,14 +39,12 @@
   box-sizing: border-box;
   cursor: pointer;
   flex-shrink: 0;
-  height: common.$header-height;
   padding: common.$spacing-medium common.$spacing-xl;
 }
 
 .nav-item-wrapper {
   @include common.flex-row-align-center;
   gap: common.$spacing-small;
-  height: common.$header-height;
 }
 
 .nav-item {
