+++ lit/src/sass/_common.scss
@@ -166,7 +166,6 @@ $spacing-large: 12px;
 $spacing-xl: 16px;
 $spacing-xxl: 24px;
 
-$header-height: 44px;
 $footer-height: 60px;
 $chat-height: calc(100vh - 320px);
 $main-content-max-width: 1500px;
